rizzler_client/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dearpygui==2.0.0
docstring-parser==0.16
executing==2.2.0
filetype==1.2.0
frozenlist==1.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.18.2
google-crc32c==1.7.1
google-genai==1.0.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx<0.28
httpx-sse==0.4.0
icecream==2.1.4
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.3.27
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
marshmallow==3.26.1
mcp==1.6.0
mcp-server-fetch==2025.4.7
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.5
ollama==0.4.7
orjson==3.10.16
packaging==24.2
propcache==0.3.1
protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.1.2
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vertexai==1.71.1
webencodings==0.5.1
websockets>13.0
yarl==1.20.0
zstandard==0.23.0
